+++ NEWS.md
@@ -156,3 +156,9 @@ First set of updates for simulating with recombination!
 # simfam 1.1.9.9000 (2024-07-17)
 
 - Function `pop_recomb` added support for option `haps` to be a `BEDMatrix` object, and added option `loci_on_cols` to accept a transposed `haps` input.
+
+# simfam 1.1.10.9000 (2024-10-22)
+
+- Function `pop_recomb` added option `indexes_loci` to simulate only a portion of the genome available.
+- Added functions `tidy_recomb_map_inds` and `recomb_founder_blocks_inherited` to create and manipulate tidy versions of our recombination data structure, which are more useful in some cases, currently focused on identifying founder blocks that are inherited by focal individuals.
+  - Added `dplyr` and `tidyselect` as new package dependencies, which is used for both of these new functions that manipulate tidy tables.